update minor rev to 0.2.1

+++ powerprotect/protectionrule.py
@@ -1,3 +1,6 @@
+from powerprotect.ppdm import ReturnValue
+
+
 class ProtectionRule:
 
     def __init__(self, **kwargs):
@@ -22,11 +25,15 @@ class ProtectionRule:
 
     def delete_rule(self):
         if self.exists:
-            return_value = self.ppdm.delete_protection_rule(self.body['id'])
+            if not self.check_mode:
+                return_value = self.ppdm.delete_protection_rule(self.body['id'])
+                self.exists = False
+            if self.check_mode:
+                return_value = ReturnValue()
+                return_value.success = True
             if return_value.success:
                 self.changed = True
                 self.body = {}
-                self.exists = False
                 self.msg = f"Protection rule {self.name} deleted"
             elif return_value.success is False:
                 self.failure = True
@@ -37,13 +44,17 @@ class ProtectionRule:
         inventory_type = kwargs['inventory_type']
         label = kwargs['label']
         if not self.exists:
-            return_value = self.ppdm.create_protection_rule(rule_name=self.name,
-                                                       policy_name=policy_name,
-                                                       inventory_type=inventory_type,
-                                                       label=label)
+            if not self.check_mode:
+                return_value = self.ppdm.create_protection_rule(rule_name=self.name,
+                                                                policy_name=policy_name,
+                                                                inventory_type=inventory_type,
+                                                                label=label)
+                self.get_rule()
+            if self.check_mode:
+                return_value = ReturnValue()
+                return_value.success = True
             if return_value.success:
                 self.changed = True
-                self.get_rule()
                 self.msg = f"Protection Rule {self.name} created"
             elif return_value.success is False:
                 self.failure = True
@@ -53,11 +64,14 @@ class ProtectionRule:
         if (self.exists and
             self.__body_match(self.body, self.target_body) is False):
             self.body.update(self.target_body)
-            return_value = self.ppdm.update_protection_rule(self.body)
+            if not self.check_mode:
+                return_value = self.ppdm.update_protection_rule(self.body)
+                self.get_rule()
+            if self.check_mode:
+                return_value = ReturnValue()
+                return_value.success = True
             if return_value.success:
                 self.changed = True
-                self.get_rule()
-                self.created = True
                 self.target_body = {}
                 self.msg = f"Protection Rule {self.name} updated"
             elif return_value.success is False:
